scripts/requirements.txt
peewee==3.17.0
loguru==0.7.2
pytest==7.4.2
pre-commit==3.5.0
tqdm==4.66.1
psycopg2-binary==2.9.9